refactor(mma): reorganize verifier and backend sources
Group MMA verification under a dedicated directory and move CPU/CUDA
implementations into mma/backend. Update include and CUDA build paths
without changing behavior.
